% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARA_function.R
\name{ZhaoNew_Alloc}
\alias{ZhaoNew_Alloc}
\title{Allocation Function of Zhao's New Design for Binary and Continuous Response}
\usage{
ZhaoNew_Alloc(
  ptsb.X,
  ptsb.Z,
  ptsb.t,
  ptsb.Y,
  ptnow.X,
  ptnow.Z,
  response,
  omega,
  p = 0.8
)
}
\arguments{
\item{ptsb.X}{a vector of length \code{n} of the predictive covariates of previous patients. Must be binary.}

\item{ptsb.Z}{a \code{n x k}  of the prognostic covariates of previous patients. Must be binary.}

\item{ptsb.t}{a vector of length \code{n} of the treatment allocation of previous patients.}

\item{ptsb.Y}{a vector of length \code{n} of the responses of previous patients.}

\item{ptnow.X}{a binary value of the predictive covariate of the present patient.}

\item{ptnow.Z}{a vector of length \code{k} of the prognostic covariate of the present patient.}

\item{response}{the type of the response. Options are \code{"Binary"} or \code{"Cont"}.}

\item{omega}{a vector of length \code{2+k}. The weight of imbalance.}

\item{p}{a positive value between 0.75 and 0.95. The probability parameter of Efron's biased coin design.}
}
\value{
\item{prob}{Probability of assigning the upcoming patient to treatment A.}
}
\description{
Calculating the probability of assigning the upcoming patient to treatment A based on the patient's predictive covariates and the previous patients' predictive covariates and responses for Zhao's New procedure.
}
\examples{
set.seed(123)
ptsb.X = sample(c(1, -1), 400, replace = TRUE)
ptsb.Z = cbind(
 sample(c(1, -1), 400, replace = TRUE),
 sample(c(1, -1), 400, replace = TRUE)
)
ptsb.Y = sample(c(1, 0), 400, replace = TRUE)
ptsb.t = sample(c(1, 0), 400, replace = TRUE)

## Incoming patient
ptnow.X = 1
ptnow.Z = c(1, -1)

## Run allocation probability calculation
prob = ZhaoNew_Alloc(
 ptsb.X = ptsb.X,
 ptsb.Z = ptsb.Z,
 ptsb.Y = ptsb.Y,
 ptsb.t = ptsb.t,
 ptnow.X = ptnow.X,
 ptnow.Z = ptnow.Z,
 response = "Binary",
 omega = rep(0.25, 4),
 p = 0.8
)

## View allocation probability for treatment A
prob
}
